dvfs: Extract dvfs_kfile.c from dvfs.c
