+++ src/fs/dvfs/Mybuild
@@ -23,6 +23,7 @@ module core extends embox.fs.fs_api {
 	option boolean use_dcache=false
 
 	source "dvfs.c"
+	source "dvfs_kfile.c"
 	source "dvfs_dentry.c"
 	depends embox.fs.dvfs_idesc_ops
 	source "dvfs_util.c"
